requirements.txt
fastapi
uvicorn
langchain
pymongo
chromadb
google-generativeai
python-dotenv
PyPDF2
langchain-community
langchain-google-genai
python-multipart